requirements.txt
python==3.9.13
pytorch==1.11.0
torchvision==0.12.0
numpy==1.24.3
tqdm==4.66.3
timm==0.6.13
scipy==1.10.1
opencv-python
scikit-image
